Refactor stacking. Stacks are now part of Item Data, and stack instanced structs, take care of most stack behavior as per user implementation.
Provided defaults for No Stacking (always creates new instance), Unique (only item of given Item Definition can be instanced) and Stacking (where items stack up to Max Stacks and then create new instance).

+++ ArcX/ArcCore/Source/ArcCore/Items/ArcItemSpec.h
@@ -230,7 +230,7 @@ public:
 	FArcItemId ItemId;
 
 	UPROPERTY(EditAnywhere, Category = "Data")
-	int32 Amount = 1;
+	uint16 Amount = 1;
 
 	UPROPERTY(EditAnywhere, Category = "Data")
 	uint8 Level = 1;
